rename CreateSpecContent
