+++ packages/app/src/specs/DefatultPatternNoSpecContent.vue
